Support online configuration
Configuration is done now via a Python module with key value pairs, and
you can reload the configuration while the mount is online.

The intended use is to start without any timeouts for setting up the
server, and when the system ready for testing a slow storage, modify the
configuration and reload.